nrf_security: Avoid enabling NRF_SECURITY by default
when the option GEN_ISR_TABLES is not enabled. NRF_SECURITY
for the nRF54L and nRF71 series uses the PSA CRACEN driver
which uses interrupts during its initialization. If the
GEN_ISR_TABLES is not enabled the user must manually configure
the CRACEN interrupts in order to boot the application. It is
better in such a case to require for the user to manually
enable NRF_SECURITY so that they can identify more easily
what caused the booting issue.

The GEN_ISR_TABLES is only disabled in one single Zephyr test
at the time of writing and it not widely used at all so this
should have very minimal impact.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>