west_commands: ncs_ironside_se_update: improve robustness of script
Add various improvements to the west command used to update IronSide SE:

The script now updates both the main and recovery firmwares by default,
unless you explictly tell it to update just one slot using
--firmware-slot <slot>. This new default behavior should be more
intuitive and fit what most users of the script want.

The script now also reads the target firmware version from each
firmware DFU blob and compares it against the firmware read from the
device after the update, erroring out if the update was not successful.
A descriptive error message is output based on the last update status.

Instead of waiting for a fixed time for the device to boot after a
reset, the script now polls the boot status using nrfutil device
x-boot-mode-get and continues when it sees a non-zero boot status.
This should make the script a bit quicker.

The --serial argument is now actually optional like the CLI indicated.

Some of the log messages have been rewritten to try to make them a bit
clearer.

Ref: NCSDK-34614

Signed-off-by: Jonathan Nilsen <jonathan.nilsen@nordicsemi.no>