modules: tf-m: group TFM_CMAKE_OPTIONS
Group most of the appending to TFM_CMAKE_OPTIONS in this CMakeLists.txt
file into a single set_property() call.

This allows to reduce the number of lines and to make it more obvious
that all we are doing is passing configuration options to TF-M.

Some occurrences are left untouched. They use more complex conditions
than just a single variable boolean check, which would become unreadable
with generator expressions.

Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>