applications: machinelearning: Remove thingy 53 non secure target
Remove the thingy53 non secure target from the machine learning
application since it enables BT_CRYPTO.

Recently the configuration of TF-M was updated to take into
account the BT_CRYPTO symbol. Meaning that when the BT_CRYPTO
is enabled TF-M configuration makes sure to provide support
for the BT_CRYPTO requirements. This was not happening before
resulting to invalid configurations where the samples failed
at runtime.

The thingy 53 non secure target has a flash overflow when trying
to build this application with a valid TF-M configuration and thus
I removed it from the yaml file.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>