applications: nrf_desktop: update documentation for nRF54LM20 DK target
Updated the nRF Desktop documentation for the nRF54LM20 DK target to
align it with the newest configuration changes. The nRF54LM20 DK target
now uses the KMU to store the MCUboot public key for image verification
and hardware-based cryptography based on the CRACEN HW peripheral. The
hardware cryptography is enabled for both the MCUboot bootloader and
the application image.

Ref: NCSDK-34043

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>