modules: tf-m: Fix mcuboot_single_image bug
The cmake logic which checks the mcuboot_single_image
in order to enable the NRF_NS_SECONDARY variable is wrong
since it evaluates the actual mcuboot_single_image
string and not the variable. As a result the boolean
check is always true and thus the NRF_NS_SECONDARY
is never enabled. Fix the check to use the variable
and not the literal string.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>