bluetooth: cs_de: Changes to IFFT algo
The IFFT distance estimate algorithm should be able to handle missing
channels.
This commit removes the assumption that it would not.
- Remove interpolation of IQ values for channels 23,24,25. Just set the
  IQs for these channels to 0.
- Remove early return in `calculate_dist_ifft` when IQ values for one
  channel would be missing. The algo should handle this ok and we can
  still produce a distance estimate.

Another change included in this commit is to resort to not doing any
"left null compensation" in the specific case where:
- The left_null_index is greater than the peak_index but not large
  enough for the left null compensation to result in a positive distance
  estimate.

Previously this case would result in the distance estimate of 0.

Signed-off-by: Erik Sandgren <erik.sandgren@nordicsemi.no>