applications: nrf_desktop: nrf54h20dk: use MCUboot custom code for PM
Enabled the CONFIG_SOC_NRF54H20_PM_S2RAM_OVERRIDE Kconfig option in
the MCUboot image configuration for the nRF54H20 DK target. This change
is aligned with the general direction of the MCUboot-based DFU solution
for the nRF54H20 SoC. It limits the bootloader reponsibility for the
wake-up from the Suspend To RAM (S2RAM) state to the bare minimum and
allows more flexibility when it comes to future improvements for the
power management system. These improvements can be easily updated with
the DFU in the application image. On the other hand, the PM code cannot
be changed in the MCUboot image in the most common case where the
MCUboot is set up as an immutable bootloader.

Ref: NCSDK-35408

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>