lib: pdn: Fix handling of APN rate control and malformed +CGEV
+CGEV: APNRATECTRL STAT can also have status 2, which means that
no APN rate control is configured for the CID. Fixed the
implementation to map this to PDN_EVENT_APN_RATE_CONTROL_OFF event
instead of PDN_EVENT_APN_RATE_CONTROL_ON.

Fixed handling of malformed +CGEV notifications where the CID can
not be parsed. Earlier these caused the malformed event to be sent
for all PDNs. Only PDN_EVENT_NETWORK_DETACH is intended to be sent
for all PDNs, because the AT notification does not contain a CID
and it affects all PDNs.

Signed-off-by: Tommi Kangas <tommi.kangas@nordicsemi.no>