nrf_security: cracen: Pull out mbedtls_psa_platform_get_builtin_key
If multiple crypto drivers are used and they all define built-in keys,
then this platform function needs to account for all of these drivers.
Therefore, move it into a new top-level file: `mbedtls_psa_platform.c`.

The new file is placed next to `psa_crypto_driver_wrappers.c`, to which
it is closely related.

Ref: NCSDK-35399
Signed-off-by: Grzegorz Swiderski <grzegorz.swiderski@nordicsemi.no>