applications: nrf_desktop: nrf54h20dk: update memory map to optimize DFU
Updated the memory map for the nRF54H20 DK in the nRF Desktop
application to optimize the size of the DFU package. This change
optimizes the size of the DFU binary by making it around 256 kB
smaller. As a result, the DFU transfer is faster than with the
previous configuration.

Ref: NCSDK-35817

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>