soc: nrf71: Fix false positive compile array index OOB warning
The SPU ID calculation is correct and is never OOB (<=63) but if we keep
the function public then compiler finds a false positive case where
index is 64. I have verified that this is not the case for any
peropheral, but unsure why compiler warns.

Adding an unreachable check solves but might miss actual error cases.
Adding static is simpler and limits the scope to fix false positive.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>