bluetooth: cs_de: Fix issue in IFFT calculation
This commit fixes an issue with how cs_de implemented IFFT, the issue
caused a 1 bin offset in the resulting IFFT magnitude, which resulted
in a constant offset of distance estimations by 1 bin.

The issue was how IFFT was implemented using the FFT, the result of FFT
should be reversed but this was done in an incorrect way (the 0th
element should be fixed) whereas all other elements should be reversed
order.

The fix however chooses a different approach to find the IFFT
using the FFT:
1. Complex conjugate the input.
2. Perform the FFT.
3. Complex conjugate the output.

Since we are interested in the magnitude of the IFFT,
we skip step 3. and directly calculate
the magnitude of the output of step 2.

The new approach is tested and verified to be equal to the
output of numpy IFFT:
`np.abs(np.fft.ifft(iq_tones_comb, n=512))`

Signed-off-by: Erik Sandgren <erik.sandgren@nordicsemi.no>