bootloader: Fix NSIB on nrf54l15dk/nrf54l15/cpuapp/ns
This commit fixes the device not booting on
nrf54l15dk/nrf54l15/cpuapp/ns target when NSIB is used.

The issue was being a result of the SRAM used by B0
not being properly placed (at sram_nonsecure, while
it should be at sram_secure). This resulted in KMU
hardware errors.

Signed-off-by: Artur Hadasz <artur.hadasz@nordicsemi.no>