mpsl: hwres: dppi: Adapt ppib_get_domain for bsim
We need to update the changes done in
00bd43cf86f94389bbcaebfe70e563eb503b9532 to support compiling
the code for bsim. A switch case statement cannot be used
because it requires compile time constants in each label.
When compiling for babblesim this is not the case.

Simply transforming it to a set of if statements also works.

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>
(cherry picked from commit 0d181b5cc5260585982fa9240d376bd73d4b6b6d)