samples: bluetooth: stabilize iso time sync sample
Stabilize the ISO time sync sample for nRF53 for the use of the sample
with the ipc_radio on the network core.

In our testing, we observed that the transport of data from the
application to the network core takes slightly longer. To avoid
flushed packets, I increased the time margin to send SDUs to the
network core slightly earlier.

Signed-off-by: Jan Müller <jan.mueller@nordicsemi.no>